[release-v1.7] build: Enable run_tests.sh to work with go.work.
The script `run_tests.sh` does not work if a go.work file is present in
the project root. This is because when a `go.work` file is present, `go
list -m` returns the names of all modules in the workspace rather than
just the parent module in the root directory. Removing the `-m` flag
fixes this.