rpcserver: Use atomic for ws client disconnect.
This modifies the logic that detects when a websocket client has been
disconnected to make use of a more efficient atomic int instead of a
bool gated by a mutex and uses the atomic as the condition for exiting
the input handler versus relying on a less efficient non-blocking select
on the quit chan.

It also avoids logging an error when the input handler fails to read a
message due to the client being disconnected.  This prevents the
spurious "websocket: close 1006 (abnormal closure): unexpected EOF"
error messages when websocket clients disconnect.