rpcserver: Consistent block connected ntfn skip.
This updates the websocket notification manager block connected handler
to skip creation of the associated notification when no clients have
requested it inside of the handler instead of prior to it so it is
consistent with the rest of the handlers.

It was previously done outside of the handler in order to skip a bunch
of work that is no longer relevant due to the change to the filters some
time ago.