rpcserver: Pass request context to handlers.
Currently, the handler for the HTTP endpoint passes the server context
to the handlers which ensures that they can cancel immediately in the
event of server shutdown.  Meanwhile, the WebSocket endpoint relies on
the server disconnecting the clients when it is shutdown in order to
signal cancellation.  In both cases, however, the context is not
cancelled when a client initiates the disconnect which means handlers
do not have the opportunity to cancel early in that case.

In order to make it so that the handlers can cancel early in all of the
aforementioned scenarios, this modifies the handlers for both the HTTP
and WebSocket endpoints to make use of the request context instead of
the server context and arranges for each request context to use the
server context as their parent context.

It also modifies all of the individual websocket-specific handlers to
accept the context in the same way the shared handlers for both
endpoints do so they are able to make use of early cancellation if
desired as well.