rpcserver: Support getwork cancellation.
Currently, a mutex is used in order to protect concurrent access from
multiple RPC invocations for work requests and submission.  While this
approach achieves the desired result and is logically correct, it also
means that any requests that are cancelled needlessly block those that
haven't been cancelled.

This is particularly noteworthy around initial startup where it can take
a while to reach the initial state sync and provide a usable template
and thus any callers, such as gominer, that have timeouts involved will
cancel request and make new ones.

In order to better handle cases such as the aforementioned, this
modifies the getwork handler to support cancellation in both the
concurrent access blocking behavior as well as when waiting for template
generation.

It also reduces the time spent in the critical section of the work state
mutex which can help with the throughput of work notifications in high
contention scenarios.