rpcserver: Detect disconnect of ws clients.
The recent commit that converted the websocket handler to use contexts
failed to account for the request context not being closed when the
connection is closed unlike the case in the HTTP path since the
websocket hijacks the connection.

This ensures the Run method terminates on websocket disconnection by
selecting across both the context and the quit channel that is closed
when disconnect is called.