docs: Make JSON-RPC rescan docs match reality.
The semantics for discovering transactions that involve specific
addresses and outpoints that a client is interested in changed some time
ago to be based on loading transaction filters.

It appears that the JSON-RPC API documentation was not updated to match
the new reality at that time, so this updates the documentation
accordingly.

In particular, it updates the JSON-RPC API documentation for the
loadtxfilter, notifynewtransactions, and rescan methods as well as the
blockconnected and blockdisconnected notifications to match the correct
parameters and semantics.

Next, since rescan now blocks until it finishes and therefore no
longer is involved with or sends notifications, this removes references
to rescan from all notifications and removes the longer available
rescanprogress and rescanfinished notifications.

Finally, it adds documentation for the releveanttxaccepted notification
sent in response to a transaction that matches the loaded transaction
filter being added to the mempool.