main: Remove old style build constraints.
This removes the old style build constraints that were required for
versions of Go prior to version 1.17 now that now that the minimum
actively supported version of Go is 1.18.

Note that it intentionally does not remove them from the modules to
avoid needlessly breaking any consumers that might still be using them
from older versions of Go.