mempool: Accept expired prune height.
This modifies the method that prunes expired transactions to accept the
current best height as opposed to looking it up internally.

While this change doesn't really have any effect in isolation, it is
part of an ongoing effort to make the internal mempool state easier to
reason about for correctness by removing hidden assumptions.  In this
particular case, it removes the reliance on the assumption that the tip
of the chain has not changed as compared to what the caller expects it
to be.