chaincfg: Rework deployment validation tests.
This reworks the deployment validation tests to make them more
consistent with modern practices in the code as well as to simplify them
and make them more flexible by moving the primary logic for all of the
expected results into the test definitions themselves.

It also tests for the missing abstain and missing no choice cases.