chaincfg: Enforce globally unique vote IDs.
The code currently supports reusing the same vote ID to refer to
different agendas as long as they are in different agenda versions.

While there is nothing technically incorrect about this, and part of the
design is to allow it, various other software in the ecosystem
implicitly assumes the IDs are globally unique across all agendas.

Due to this, and also because supporting it makes the internal code more
complex to deal with as well, there is currently a bit of an unwritten
rule that the vote ID for every agenda across all versions should be
unique.

This change effectively formalizes that unwritten rule by modifying the
validation logic that happens at initialization time to reject
deployments that have duplicate IDs across agenda versions.