blockchain: Simplify threshold state determination.
This simplifies the internal logic for determining threshold states for
various deployments by introducing a structure to house additional
information about a deployment and passing that around as opposed to
previous approach of passing around the information needed to lookup the
relevant information and associated caches.

The individual structs are created from the chain parameters and stored
in a map when the chain instance is initialized.