multi: Add forced deployment result network param.
Currently, the lack of an entry for a given deployment in the chain
parameters for a given network is treated the same as the deployment
being active.  This behavior is primarily in order to support testing
without needing to go through the entire voting process on testing
networks such as the simulation network as well as allows rule changes
on newer versions of testnet networks that were previously accepted on
prior versions.

However, the existing approach has been pointed out on several occasions
as being surprising behavior by those working on new consensus changes
and it also means testing the voting process itself for new rule changes
prior to their release requires a cumbersome sequence of not so obvious
steps.  Further, it also makes those historical pruned rule changes
invisible to RPCs like getblockchaininfo.

In order to address the aforementioned inconveniences, this adds a new
parameter to the per-network chain parameters to specify a vote choice
that should always be considered as having been the majority result of a
vote and updates the relevant logic in blockchain accordingly to respect
it.

The new approach does not suffer from the aforementioned inconveniences
as it permits a simple single line modification to allow the voting
process to be tested as well as makes automatically activated historical
rules visible to the RPCs like getblockchaininfo.

Finally, this also adds previously pruned deployment parameters back to
the relevant network parameters with the new field set to force them to
be active.