apbf: Prepare v1.0.1.
This updates the apbf module to use the latest siphash module
dependency, the copyright year in the files modifies since the previous
release, and serves as a base for container/apbf/v1.0.1.

The updated direct dependencies in this commit are as follows:

- github.com/dchest/siphash@v1.2.3

The full list of updated direct dependencies since the previous
container/apbf/v1.0.0 release are the same as above.

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.