blockchain: Remove unused db update tracking.
Many years ago, the various threshold states were stored in the database
due to performance reasons.  However, that logic was removed when the
code was updated to support voting since other changes made it quite
efficient to simply calculate the states at startup.

However, it appears that the dbUpdates field was not removed even though
it's not actually used anymore.  This removes the field and associated
logic accordingly.