blockchain: Validate deployment chain params.
This adds code to the chain initialization process to validate the
deployment parameters defined in the given chain params satisfy the
assumptions that the vote tallying logic rely on.

It also adds a test for the new validation logic along with the existing
validation items that were not already directly tested.

This is similar to the recently-added validation logic in the chaincfg
module itself, except it focuses specifically on the subset the
blockchain code requires which provides a path for optimizing and
simplifying the code related to vote tallying.