blockchain: Remove deployment checker abstraction.
The existing threshold state determination code was written long ago
prior to the addition of the decentralized rule change consensus voting
semantics with the intention of it providing an abstract way of
implementing generalized threshold states via a deployment checker
abstraction.

However, when the consensus rule voting logic was added, it became clear
that it wasn't sufficiently abstract enough to support everything that
was required which led to the implementation of the vote tallying logic
being somewhat inefficient and harder to reason about than is ideal.

Consequently, this reworks the threshold state determination and vote
tallying logic to remove that abstraction while simultaneously
optimizing it and simplifying the implementation to make it easier to
reason about for correctness.