chainhash: Add function to return a hash.Hash
hash.Hash allows the data being hashed to be written to incrementally,
as it implements io.Writer.  This allows large byte sequences to be
hashed without needing a single large byte slice in memory.  While the
hash.Hash object requires an allocation, the trade-off of not needing
the entire byte sequence in a single byte slice may be worthwhile.