blockchain: Reject params with mask approval bit.
This adds an additional validation check to the chain initialization
process to validate that the mask of the deployments in the given chain
params do not use the bit that is reserved for use to signal whether or
not the regular transaction tree of the parent block should be
considered valid.

It also adds a test for the new validation logic.