netsync: Re-request data sooner after peer disconnect.
This change speeds up the relay of blocks and transactions when the
peer we have queried for this data disconnects.  If any other peers
have also announced this data, request the data again immediately from
them instead of waiting for a slower peer to announce the data again
through an inventory message.