blockchain: Convert several test helpers.
This converts several of the funcs used in the tests that serve to find
deployments, votes, and choices within parameters to fully fledged test
helpers that invoke fatal upon error instead of returning the errors
themselves.

This makes them more convenient to use and makes the tests slightly more
compact in general as well.