blockchain: Use local mock votes in tests.
This modifies the tests to make use of a couple of functions to acquire
mock votes instead of global ones.  This helps reduce global state in
the tests and makes it much harder to inadvertently break test
assumptions when they are modified since each test now has their own
copy.

It also now uses the test helper funcs to look up the choices within the
votes to avoid hard coded magic numbers.