blockchain: Reassign non-overlapping test params.
This introduces a helper func to the tests that updates the masks and
choice bits for all votes in a provided test agenda so that they are
non-overlapping.

It then updates a couple of the tests that merge votes into the same
agenda to make use of the func so they no longer involve overlapping
bits.