[release-v1.7] addrmgr: Support module graph prune and lazy load.
This bumps the go directive for the addrmgr module to 1.17 which will allow
the new module graph pruning and lazy loading capabilities introduced in
Go 1.17 to be used once the updated modules are released.

This means that, as described by the documentation, the go.mod file for
the module now include a separate require block that includes all of the
indirect dependencies.