chaincfg: Introduce BLAKE3 PoW agenda.
This adds a new definition for the upcoming agenda vote to change the
proof-of-work hashing function to BLAKE3 as defined in DCP0011.

It does not include any code to make decisions based on the status of
the agenda or bump block versions.  It only adds the definition for the
agenda.