chaincfg: Introduce subsidy split change r2 agenda.
This adds a new definition for the upcoming agenda vote to change the
subsidy split to 1% PoW, 89% PoS, and 10% Treasury defined in DCP0012.

It does not include any code to make decisions based on the status of
the agenda or bump block versions.  It only adds the definition for the
agenda.