standalone: Add modified subsidy split calcs.
This adds two new exported functions to the subsidy cache in
blockchain/standalone to calculate the work and stake vote subsidies
according to a provided subsidy split variant flag along with tests to
ensure expected behavior.

The available variants are:

- The original subsidy split used at launch
- The modified subsidy split defined by DCP0010
- The modified subsidy split defined by DCP0012

It should be noted that the values for the modified split are hard coded
as opposed to using the subsidy cache params in order to avoid the need
for a major module bump that would be required if the subsidy params
interface were changed.  The values are the same for all networks, so no
additional logic is necessary on a per-network basis.