chaincfg/chainhash: Prepare v1.0.4.
This updates the chaincfg/chainhash module to use the latest blake256
module dependency and serves as a base for chaincfg/chainhash/v1.0.4.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.0.1

The full list of updated direct dependencies since the previous
chaincfg/chainhash/v1.0.3 release are the same as above.

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.