secp256k1: Prepare v4.1.1.
This updates the secp256k1 module dependencies, the copyright year in
the files modified since the previous release, and serves as a base for
dcrec/secp256k1/v4.1.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.0.1

The full list of updated direct dependencies since the previous
dcrec/secp256k1/v4.1.0 release are the same as above.

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.