dcrjson: Prepare v4.0.1.
This updates the dcrjson module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
dcrjson/v4.0.1.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.4

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.0.1

The full list of updated direct and indirect dependencies since the
previous dcrjson/v4.0.0 release are the same as above.

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.