rpcserver: Use solved mock orphan block in tests.
Currently, the RPC server getwork test for orphan block submission is
rather brittle because it relies on the specific current implementation
of the handler.

This makes the test more robust by modifying it to solve the mock orphan
block submission so that it reliably tests what it intends to even if
the handler code changes.