rpcserver: Consolidate mock mining addr.
This modifies the RPC server tests to create a mining address as a part
of the default mining state and use that state throughout.  It also
removes the separate mock mining address override field since the same
thing can be accomplished via the mock mining state now.

This has the effect of consolidating the logic and simplifying its
usage.