blockchain: Correct total subsidy calculation.
This modifies calculateAddedSubsidy to also include the subsidy
generated by treasurybases to ensure the total reported coin supply is
the expected value.

It should be noted that this commit only corrects the function and the
issue does not affect consensus in any way since the calculated value is
only used when providing the information via RPC.

However, it does mean the database has an incorrect value stored.  That
will be corrected via a database migration in a separate commit.