blockchain: Correct total subsidy database entry.
This adds code to correct the total subsidy entry in the database due to
the issue in calculateAddedSubsidy described by the previous commit.

It involves loading and iterating through all of the main chain blocks
and to recalculate the total subsidy and then updating the best chain
state that houses the stored value.  The process can be interrupted at
any point and future invocations will resume from the point it was
interrupted.

Also, this takes advantage of the database version bump to only check
and remove  the legacy spend consumer dependencies bucket once during
the upgrade instead of at every startup.