chaingen: Add blake3 support.
This adds support for generating blocks solved by blake3 to the test
chain generator.

It accomplishes this by introducing a new type and method, named
PowHashAlgorithm and UsePowHashAlgo, respectively, to the chain
generator along with logic to change the hash function accordingly.  The
method takes a parameter of the new type in order to allow the caller
selectively choose which hash function to use.

This will allow future code to test all edge conditions around switching
the hash function over to blake3 as defined in DCP0011.