standalone: Add separate proof of work hash check.
This adds a new function to the blockchain/standalone module named
CheckProofOfWorkHash which can be used to independently verify that a
provided hash is less than a provided compact target difficulty without
also additionally checking that the compact difficulty is within a valid
range.

Since there is already CheckProofOfWorkRange to perform the latter, this
provides callers additional flexibility of performing the required
checks independently without being forced to re-check the range.

It also updates includes comprehensive tests.