rpcserver: Consolidate work data serialization.
This consolidates the logic that serializes the data that represents
work to be solved in the getwork RPC and notifywork WebSocket
notification.

The work data consists of the serialized block header along with any
internal padding that makes the data ready for callers to make use of
only the final chunk along with the midstate for the rest when solving
the block.

Having it serialized in one place is both more convenient and less error
prone since changing it will now automatically apply to all methods of
obtaining work.

It also updates tests to use the new function.