standalone: Add ASERT difficulty calculation.
This adds a new exported function to the blockchain/standalone module
named CalcASERTDiff which can be used to calculate a required target
difficulty using the Absolutely Scheduled Exponentially weighted Rising
Targets (ASERT) algorithm defined in DCP0011.

It also updates the documentation and includes comprehensive tests.

Note that this commit only makes the new function available and
does not implement any logic based on it.