chaingen: Add ASERT difficulty algorithm support.
This adds support for generating blocks that require the new difficulty
algorithm defined in DCP0011 to the test chain generator.

It accomplishes this by introducing a new type and method, named
PowDifficultyAlgorithm and UsePowDiffAlgo, respectively, to the chain
generator along with logic to calculate the appropriate difficulty
accordingly.  The method takes a parameter of the new type in order to
allow the caller selectively choose which difficulty algorithm to use.

This will allow future code to test all edge conditions around switching
the difficulty algorithm over to the new one defined in DCP0011.