wire: Prepare v1.6.0.
This updates the wire module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
wire/v1.6.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.4

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.0.1

The full list of updated direct dependencies since the previous
wire/v1.5.0 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.4
- lukechampine.com/blake3@v1.2.1

The full list of updated indirect dependencies since the previous
wire/v1.5.0 release are as follows:

- github.com/decred/dcrd/crypto/blake256 v1.0.1
- github.com/klauspost/cpuid/v2@v2.0.9

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.