blockchain/standalone: Prepare v2.2.0.
This updates the module dependencies, the copyright year in the files
modified since the previous release, and serves as a base for
blockchain/standalone/v2.2.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@1.0.4
- github.com/decred/dcrd/wire@v1.6.0

The updated indirect dependencies in this commit are as follows:

- github.com/decred/crypto/blake256@v1.0.1
- github.com/klauspost/cpuid/v2@v2.0.9
- lukechampine.com/blake3@v1.2.1

The full list of updated direct and indirect dependencies since the
previous blockchain/standalone/v2.1.0 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.