addrmgr: Prepare v2.0.2.
This updates the addrmgr module dependencies and serves as a base for
addrmgr/v2.0.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.4
- github.com/decred/dcrd/wire@v1.6.0

The updated indirect dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.0.1
- github.com/klauspost/cpuid/v2@v2.0.9
- lukechampine.com/blake3 v1.2.1

The full list of updated direct and indirect dependencies since the
previous addrmgr/v2.0.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.