rpcclient: Prepare v8.0.0.
This updates the rpcclient module dependencies and serves as a base for
rpcclient/v8.0.0.

The full list of updated direct dependencies since the previous
rpcclient/v7.0.0 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.4
- github.com/decred/dcrd/dcrjson/v4@v4.0.1
- github.com/decred/dcrd/dcrutil/v4@v4.0.1
- github.com/decred/dcrd/gcs/v4@v4.0.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v4@v4.0.0
- github.com/decred/dcrd/txscript/v4@v4.1.0
- github.com/decred/dcrd/wire@v1.6.0

The full list of updated indirect dependencies since the previous
rpcclient/v7.0.0 release are as follows:

- github.com/agl/ed25519@v0.0.0-20170116200512-5312a6153412
- github.com/dchest/siphash@v1.2.3
- github.com/decred/base58@v1.0.5
- github.com/decred/dcrd/blockchain/stake/v5@v5.0.0
- github.com/decred/dcrd/chaincfg/v3@v3.2.0
- github.com/decred/dcrd/crypto/blake256@v1.0.1
- github.com/decred/dcrd/crypto/ripemd160@v1.0.2
- github.com/decred/dcrd/database/v3@v3.0.1
- github.com/decred/dcrd/dcrec@v1.0.1
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.3
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.2.0
- github.com/klauspost/cpuid/v2@v2.0.9
- lukechampine.com/blake3@v1.2.1

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.