server: Force PoW upgrade to v10.
This modifies the vote notification logic so that wallets will no longer
vote on mainnet blocks once height 777240 has been reached and the block
version is prior to 10.

This change is being made to make use of the staking system to force
proof-of-work miners to upgrade to the latest version in a timely
fashion so voting on the new consensus changes can commence.