build: Add CI support for test and module cache.
This makes use of the github actions cache to save and restore the Go
test and module caches.  This should result in faster CI runs since most
changes only affect a small number of tests.

It also updates the RPC integration tests to ensure the logs directory
is not in the source code dir so the go testing cache can be used.

Next, it renames the txscript module's data dir to testdata so go does
not treat it like a package.

Finally, it adds a script that stabilizes the timestamps on all files in
testdata directories when running the github action.  This is necessary
because the go testing cache logic uses the timestamps of all input
files when determining whether or not a test needs to be rerun and
github clones a fresh repo on every run which causes the timestamps of
the test data to change.