database: Use TempDir to create temp test dirs.
This modifies several of the tests to use t.TempDir.  The directory it
creates is automatically removed when the test and all its subtests
complete.

It also changes out some of the cleanup logic for closing the database
to use t.Cleanup instead of defer to ensure the close happens during
test cleanup after the directories created by t.TempDir have been
removed.