config: Support DCRD_APPDATA env variable.
This modifies the config parsing to use the DCRD_APPDATA environment
variable to override the overall application data directory as an
alternative to the --appdata CLI option.

Note that the --appdata CLI option takes precedence over the environment
variable if both are specified.

Environment variables are generally the preferred approach for
overriding defaults in containers.