build: Use go install for linter and add cache.
This modifies the GitHub Action workflow to install golangci-lint from
source with go install instead of using the separate install script and
adds ~/.cache/golangci-lint to the saved cache for faster future runs.

The goal is to pin the dependency for the same reason the GitHub actions
are pinned by hash.  Namely, it reduces potential security risks such as
compromised dependencies and dependency substitution attacks.

Using go install serves to pin the dependency because go verifies the
downloaded module contents against the original checksum hashes they
were first created with.