multi: Avoid range capture for Go 1.22 changes.
Go 1.22 is introducing a change to the way the range statement works
that is not entirely backward compatible.  In particular, the loop
variable is changing from one instance per loop to one instance per
iteration.

The existing semantics are well known and in order to ensure correct
behavior and avoid potential races that would otherwise result, the
current code typically opts for using ranges with an index and creating
a local that points into the array when the range variable needs to be
used as a pointer.

In addition, there are a few remaining cases that use the common
alternative approach of capturing the range variable through self
assignment.

Both approaches will work correctly when compiled with Go 1.22 as well
as older version of Go so long as the module versions are not bumped.

However, once the modules are updated to allow support of all features
introduced by go 1.22, the second approach of capturing the range
variable would no longer be needed and as a result would very likely
need to be removed to avoid vet/linter issues.  Unfortunately, the
consequence of that change would mean building the code with older
versions of Go would start to produce incorrect code.

The first approach of using a range index that is predominantly used
throughout the code does not have that potential pitfall as it works
equally well for all of the aforementioned cases.

Thus, in order to avoid any potential issues before they ever even have
a chance to arise, this updates the few remaining instances of the
second approach to the first one or to otherwise rework the code to
avoid the need altogether.